Fix: can't checkin instances more than 65535 times
Pool uses an IO.pipe for notifications and timeouts. Whenever an
instance is available, a byte is written to the pipe. A bug occurred
on the checkout part, which never read from the pipe as long as
instances are available. It eventually filled the pipe buffer in the
kernel, and writes started to fail with Errno::EAGAIN, which
triggers the event loop, wating for the pipe to be writable, thus
blocking the pool forever.